all methods

+++ android/src/main/kotlin/com/afterlogic/crypto_plugin/crypto_stream/StreamSink.java
@@ -1,5 +1,5 @@
 package com.afterlogic.crypto_plugin.crypto_stream;
 
 public abstract class StreamSink {
-    public abstract void add(int[] bytes);
+    public abstract void add(byte[] bytes);
 }
